🎨 Use fixture and refactor error messages
